api-etl-pipeline/requirements.txt
pandas==2.1.0
requests==2.31.0
apache-airflow==2.7.2
psycopg2-binary==2.9.7
